exchange Decimal for float to improve compatibility

+++ signal_utils/common/generate_bpsk.py
@@ -1,13 +1,12 @@
 # BPSK generation - returns a complex IQ file in the range [-1 + 0j, 1 + 0j]
 
 import math
-from decimal import Decimal
 
 import numpy as np
 from numpy.typing import NDArray
 
 
-def generate_bpsk(data: NDArray[np.int_], sample_rate: int, bit_length: Decimal) -> NDArray[np.complex_]:
+def generate_bpsk(data: NDArray[np.int_], sample_rate: int, bit_length: float) -> NDArray[np.complex_]:
 
     num_bits = data.shape[0]
     samples_per_bit = math.floor(sample_rate * bit_length)
